Properly cache production bundles
To add the app to the static cache in a production build we need to know
the names. Since they are auto-generated this is not a trivial task.
Fortunately, there is a webpack plugin called
`serviceworker-webpack-plugin` that takes care of that and is not
intrusive at the same time.

We just need to pass it our service worker source file (which thus has
to move from the public folder to the src folder) and trigger the
registration through that plugin. Then we have a new struct
`serviceWorkerOption` available inside the ServiceWorker that contains a
list of static assets.