Add deployment steps to gitlab ci config
We add two new jobs: build and publish

build will run on every HEAD of every push. This is to make sure that
the current state of a merge request can be built.

publish will ultimately only run on master and produce a docker image
and push it into the gitlab registry.