Load the vapid keys and session data from a volume
The sessionsData will no longer be included in the docker image.
Instead it is mapped into the docker container via a volume. This allows
us to persist the session key outside of the container on our server and
it won't be regenerated on every rebuild of the swec-app docker image.

This also enables us to leverage the existing change detection for the
sessions data and send push notifications on changes. With the sessions
data incorporated into the docker image, the data would have to be
changed inside the git repo. Deploying it to the server will also
restart the server it self, nullifying the change detection.