Use the PushAPI-Saga connection for toast notifications
We now generate the toast on the client side based on the diff the
server sent us. Toasts are only shown for sessions we liked and the
session is mentioned in the toast. Also, we no longer try to fetch all
of the session data but instead apply the diff to our sessions state.